requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gradio==4.11.0
unsloth
torch>=2.0.0
transformers>=4.36.0
datasets>=2.15.0
python-dotenv==1.0.0
plotly==5.18.0
numpy>=1.24.0
ollama==0.1.6
requests==2.31.0
huggingface-hub>=0.19.0
imageio==2.31.6
trl>=0.7.10
pydantic==2.5.2
imageio[ffmpeg]